Note issue with grokking photon-hdf5 format
Namely, that there is not one place to reference for the file spec, like
a BNF tree. Instead, I have to hunt through project files. It would be
nice if this info was programmatically available too, and first class.
It would then be easier I think to write general programs that
manipulate such data. Maybe add to phconvert?