handle nested dicts in recursive_merge()
Final iteration on the recursive function approach. When colliding keys
both hold dictionaries, recursively call recursive_merge() on them.

Even though we are mutating and iterating, I don't think it should be a
problem. We are iterating through dom, which is not mutated. sub is
mutated, but not iterated through; we only access/write keys in sub. So
unless I'm mistaken somewhere, we won't run into issues with mutating a
dictionary we are iterating over.

This is not tail recursive. Can it be made tail recursive? Better, can
it be made iterative? Python not supporting tail recursion and all....