Move example code back to example file
This looks like it was accidentally cut and paste into the new
point-n-shoot file because it came after the logic I was working on. Not
too sure why I put my logic in the middle. Maybe for convenience because
it was closer to the code I was referencing? If I'm wrong we could have
issues later...