Pull loading out into dedicated function
I don't think this is a great choice of abstraction. The loader
functions each have slightly different return types. Meaning, special
handling could be needed for each type. In which case, maybe it would be
better to have a dedicated function for /converting/ each file type, and
then unify those under a generic interface? Hmmm... It's going to take
some mucking about to sort this through.