Incorporate recursive_merge into conversion process
Now it is possible to specify fragments of the Photon-HDF5 hierarchy to
splice into the data from the file, preserving any existing entries that
aren't explicitly intended to be overwritten.

Removed the option of passing a yaml file for now, since the
data_fragment arg could just be the results of loading such a file.
However, it would be convenient to add it back later for a more
streamlined process.