Use original py4cl
The new one (py4cl2) tries some fancy threading stuff to get real-time output
from python and it is a bit of a mess to capture it all for presentation
later. I could submit bug reports and help resolve things, but py4cl
does what I want so in the interest of time I'll just use that.