Get basic metadata file entry added
Probably a bit fragile. Among other things, it assumes that we apply the
metadata file to all entries in a directory that is added. This is not
always the case, and probably isn't the case most of the time. We do the
same thing with descriptions. It would be nice to have more fine-grained
control over batch conversions.