Try cloning phconvert into the build folder
Horrible idea, it means we're now depending on uncontrolled stuff like
the current master branch of phconvert for a successful build. That
said, this is being done because in order to build an executable with
pyinstaller we need to specify the inclusion of certain json files that
ship with phconvert. Since right now I'm just sketching things out, I'm
going to try pulling them straight from a cloned repo. We need to find a
way to either

    1) Remove this dependency on files that ship with phconvert
    2) In a controlled, reproducible, and versioned manner, extract the
    files from phconvert.