Specify the actual phconvert module folder as a place to look for modules
I was under the mistaken impression that Python /recursively/ searches
through the directories in the current working directory to find
modules. Since that isn't the case, we need to make sure that the actual
phconvert module can be found. I think we're just about at the point
where it is better to find a proper way of doing this with package
managers and stuff instead of directly cloning into the build directory.