Revert "Change release naming"
This reverts commit 0e4b3916126ce8a5a879959d213bbd3e4fd560db.

I didn't get the expected behavior, which was that upon pushing to
master the old releases would be replaced by releases of the same name
but built from the new source. At least, that appears to not have
happened judging by the appveyor logs. Even if it did, the release
retained the original commit hash, which is bad for traceability.