Stop recursing on tags!
Found out why I was getting duplicate release builds. Appveyor created a
release (with a tag!), and then seeing that a new tag was created on the
repo, proceeded to run a build for /that/ tag, resulting in at least a
duplication, possibly more than double build attempts if I recall.