Break things into pre-, build, and post- scripts
In preparation for multi-arch builds, I'm moving os-specific stuff to
pre- and post- scripts so that I can keep the common core the same. The
fact that I need to work through a configuration file, with literal
platform/shell specific code for any real logic, seems to be making
contortions like this necessary. It'd be nice to have some
platform-agnostic way of saying "expand this variable" and so on.
Frankly, what I want to do is write code not configs. I wouldn't mind
writing the CI pipeline "config" in python, with the os package taking
care of the platform-specific stuff. I wonder if I'm missing something,
or doing things wrong and that's why configuring appveyor is rubbing me
the wrong way...