Explicitly specify the build image
For now, just the default image per appveyor's docs. In the future, this
will be a sequence of the three different operating systems we can
support with appveyor. We make things explicit /now/ so that we can
special-case off the specific image being used with confidence that we
know the default value.