Specify image again, but for a newer one
So I checked the build logs, and guess what? Visual Studio 2019 folders
show up in the path but not 2015... hmmm... I checked the logs from when
I specified VS 2015, and different folders are in the path there. From
this I surmise that Visual Studio 2015 is /not/ the default image
anymore. So now we specify 2019. A quick check of the pre-installed
software verifies that the VS 2015 image does not come with miniconda
for python 3.8. In the future then, refer to the pre-installed software
table whenever the documentation elsewhere mentions default values?