Don't set PYTHON variable, we don't use it
Which begs the question, where is our python executable coming from? My
money is on conda, because poking around inside my local anaconda's
install folder shows that there is a python executable in there. Also,
from the looks of things, Appveyor doesn't put any python stuff in the
path. Instead, it installs it and lets you deal with setting up the path
so you can use it. I haven't confirmed this yet though.