Use temp files during conversion
As Xavier observed, phconvert will create an hdf5 file even if errors
are encountered, which can be confusing. To deal with this, try first
using a temp file for the conversion, and later moving it to the desired
location if the conversion is successful. Delete the tempfile when done,
regardless of whether the conversion was a success.