Use "except" not "finally"
We actually /don't/ want to always delete the temp file. If conversion
is successful, we move it to the final location. That means that if we
try to remove it after a successful conversion, we'll get a
file-not-found error (and we were). Now, only try removing it when
errors are signaled. But, make sure that the error is raised again so
that it can be handled elsewhere.