Open progress log as 'a' not 'a+'
We don't do any reading inside the 'with' body... Maybe we should
though. Depending on how many syscalls are involved with each option, it
might make more sense to generate our list of skippable files inside the
big 'with' that's being used for conversion. This would probably only
matter if oodles and boodles (and no less!) directories were being
converted.