Try fixing the malformed metadata issue
measurement_specs in the yml files are specified top-level. I
interpreted the yml files as mirrors of the photon-hdf5 data format,
which has measurement_specs nested inside another group. This caused
issues during validation of the photon-hdf5. So now, we make sure to nest
measurement_specs before performing the conversion. This isn't ideal.